Separated BindingStore out of the MST.
